+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added support of wildcard patterns for `kamu pull` and `kamu push` commands
+- Added `{dataset}/tail` and `/query` REST API endpoints
 
 ## [0.163.1] - 2024-03-01
 ### Fixed